Made it take cmdline params
Added ExpiryDateTime to force anyone who might use this script to inform
the recipient that the webhook has an expiry date.
Also intended for all halting problems to ask for any key in order to
continue but it doesn't seem to work properly and I can't be arsed
fixing it tonight.