Thought script was wrong... nope... just password
Started major changes to script, only to realise the account I was
testing it in had the password for the credential wrong. I swear this
happened last time... does Azure just not take the password correctly?
Other minor changes to code aesthetics and the use of
Login-AzureRmAccount instead of Add-AzureRmAccount...